S3C-35 Refactoring of bucketfile (S3 file metadata backend)
DISCLAIMER: not complete yet, and not properly tested, out for review
as-is to spot early issues.

The changes allow bucketfile to use a new API in Arsenal to
communicate with a remote leveldb database, containing sublevels for
bucket storage.

Instead of relying on the existing implementation of multilevel, it
uses client/server wrappers around a custom hack of 'multilevel'
module that allows it to work properly with sublevels, without
allowing custom functions per sublevel anymore (but we don't need
this so we're good).

It shall be compatible with the existing metadata since it still uses
the core sublevel module for the storage logic, only the RPC procotol
has changed.

NOTE: shall be used with Arsenal branch wip/S3C-35-customizeMultilevel