fix: delete marker creation / md
We did not reset the content length for multi-object delete requests when creating delete markers, meaning the delete marker stores content-length with the length of the xml, which is incorrect and causes the request to hang at least for file backend. It would likely also pose an issue for accurately tracking bytes created in Utapi.